requirements.txt
# Requirements for LED Matrix Project
# Install with: pip install -r requirements.txt

# GPIO control for Raspberry Pi
RPi.GPIO==0.7.1

# Optional: For more advanced LED control
# gpiozero==1.6.2

# Optional: For LED strip control (if you expand to NeoPixels later)
# rpi_ws281x==4.3.4
# adafruit-circuitpython-neopixel==6.3.9